pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "llm-portfolio-journal"
version = "1.0.0"
description = "A data-driven portfolio journal using LLMs to generate trading summaries from market data and social sentiment"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "LLM Portfolio Project"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
requires-python = ">=3.11"
dependencies = [
    "discord.py>=2.0.0",
    "pandas>=1.5.0",
    "yfinance>=0.1.96",
    "matplotlib>=3.6.0",
    "mplfinance>=0.12.9",
    "textblob>=0.17.1",
    "python-dotenv>=0.19.0",
    "requests>=2.28.0",
    "sqlalchemy>=1.4.0",
    "psycopg2-binary>=2.9.0",
    "supabase>=1.0.0",
    "openai>=1.0.0",
    "google-generativeai>=0.3.0",
    "tweepy>=4.14.0",
    "numpy>=1.24.0",
    "pytest>=7.0.0",
    "jupyter>=1.0.0",
    "ipykernel>=6.0.0",
    "pyarrow>=10.0.0",
]

[project.urls]
Homepage = "https://github.com/qmyhd/LLM-portfolio-project"
Repository = "https://github.com/qmyhd/LLM-portfolio-project"
Issues = "https://github.com/qmyhd/LLM-portfolio-project/issues"

[project.scripts]
generate-journal = "src.journal_generator:main"
portfolio-bot = "src.bot.bot:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
include = ["*"]
exclude = ["tests*"]

[tool.pylance]
# Ignore certain warnings for third-party libraries with dynamic imports
reportPrivateImportUsage = "none"
reportMissingImports = "warning"
reportMissingTypeStubs = "none"

[tool.pyright]
# Python version
pythonVersion = "3.11"
pythonPlatform = "All"

# Include and exclude patterns
include = ["src", "tests", "notebooks"]
exclude = [
    "**/__pycache__",
    ".venv",
    ".git",
    "data"
]

# Type checking mode (basic, standard, strict)
typeCheckingMode = "basic"

# Specific warning suppressions
reportPrivateImportUsage = "none"
reportMissingImports = "warning" 
reportMissingTypeStubs = "none"
reportUnknownMemberType = "none"
reportUnknownArgumentType = "none"
reportUnknownVariableType = "none"
reportGeneralTypeIssues = "warning"

# Suppress warnings for specific libraries
[tool.pyright.defineConstant]
PYLANCE_IGNORE_GOOGLE_AI = true
PYLANCE_IGNORE_TWEEPY = true
